chore: add confirm plan button
